Move documentation type classes to new documentation_types module
